requirements.txt
asgiref==3.8.1
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloudinary==1.42.2
dj-database-url==2.3.0
Django==5.1.5
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
git-filter-repo==2.47.0
h11==0.14.0
idna==3.10
outcome==1.3.0.post0
pillow==11.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
PySocks==1.7.1
python-dotenv==1.1.0
requests==2.32.3
selenium==4.29.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
